Completed rand_string

+++ print_random_string.c
@@ -3,7 +3,7 @@
 #include <time.h>
 
 
-char* rand_string(char *s, int size){
+char* rand_string(char *s, size_t size){
     for (int i = 0; i < size; i ++){
         s[i] = 'A' + (rand() % 26);
     }
